requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.4
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
mysqlclient==2.2.6
Pygments==2.18.0
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
rich==13.9.4
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.3.0
